% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace-package.R
\docType{package}
\name{spectrace-package}
\alias{spectrace}
\alias{spectrace-package}
\title{spectrace: Functions for importing and processing raw Spectrace output}
\description{
This package provides functions for importing the raw CSV output files from Spectrace into R and for processing this data to derive advanced quantities.
}
\keyword{internal}
