% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_calibrate_light2.R
\name{spectrace_calibrate_light2}
\alias{spectrace_calibrate_light2}
\title{Calibrate raw spectrace data}
\usage{
spectrace_calibrate_light2(lightData, cal_data = NULL)
}
\arguments{
\item{lightData}{Data frame with light data.}

\item{cal_data}{Data frame with calibration factors. Must consist of the
following columns: "serial", "lux", "410nm", "435nm", "460nm", "485nm",
"510nm", "535nm","560nm", "585nm", "610nm", "645nm", "680nm", "705nm",
"730nm", "760nm".}
}
\value{
Data frame with calibrated data. Columns >730nm are removed.
}
\description{
This function calibrates the raw light data of the spectrace sensors based
on the provided calibration data.
}
