% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_reference_spectra.R
\name{spectrace_cie_illuminants}
\alias{spectrace_cie_illuminants}
\title{Return CIE Illuminants}
\usage{
spectrace_cie_illuminants(as_spectrace = TRUE, resolution = "5nm")
}
\arguments{
\item{as_spectrace}{Logical. Should the spectra be returned as spectrace spectra?
Defualts to TRUE.}

\item{resolution}{String specifying the resolution of the output
spectrum. Can be "5nm" (default), "1nm", or "spectrace". The latter will result
in the original 14 spectrace channels.}
}
\value{
Data frame with CIE illuminants
}
\description{
Return CIE Illuminants
}
