% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_classify_clusters.R
\name{spectrace_classify_clusters}
\alias{spectrace_classify_clusters}
\title{Classify aggregated spectra}
\usage{
spectrace_classify_clusters(
  lightData,
  referenceData = NULL,
  aggregation = c("median", "mean"),
  method = c("correlation"),
  n.classes = 5,
  return.spectra = TRUE
)
}
\arguments{
\item{lightData}{(Grouped) data frame containing the (calibrated) light data. Data needs
to be in wide format (see \code{\link{spectrace_to_wide}}). Classification
will be performed on the aggregated data (e.g. median) per group.}

\item{referenceData}{(Optional) data frame containing the reference spectra. Needs to
contain the same wavelength columns as `lightData`. Must consist of unique
spectra and must contain a column named `spectrum_id` identifying each
spectrum. If not provided, the in-built reference spectra are used (see
\code{\link{spectrace_reference_spectra}}).}

\item{aggregation}{Aggregation method. Must be one of ['median', 'mean']. Defaults
to 'median'.}

\item{method}{Classification method. Must be one of ['correlation']. Defaults to
'correlation'.}

\item{n.classes}{Integer indicating the number of predicted classes per group
to include in the output. Defaults to 5.}
}
\value{
A data frame with the best 'n.classes' classifications (i.e. the
   'spectrum_id') and corresponding coefficients per group.
}
\description{
This function classifies (groups) of spectra by first aggregating the data and
then correlating the aggregated spectra to reference spectra, identifying the best
matching reference spectra.
}
