% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_cluster_encoded.R
\name{spectrace_cluster_encoded}
\alias{spectrace_cluster_encoded}
\title{Title}
\usage{
spectrace_cluster_encoded(
  lightData,
  lightData.encoded,
  referenceData,
  model,
  n.clusters,
  n.init,
  n.predict
)
}
\arguments{
\item{n.predict}{}
}
\description{
Title
}
