% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_combine_data.R
\name{spectrace_combine_data}
\alias{spectrace_combine_data}
\title{Title}
\usage{
spectrace_combine_data(lightData, activityData = NULL, batteryData = NULL)
}
\arguments{
\item{batteryData}{}
}
\description{
Title
}
