% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_find_sleep.R
\name{spectrace_find_sleep}
\alias{spectrace_find_sleep}
\title{Title}
\usage{
spectrace_find_sleep(
  lightData,
  actigraphyVar,
  flag_only = FALSE,
  min_length = "2 hours",
  max_interrupt = "60 mins",
  smooth_window = "10 mins",
  light_threshold = 10
)
}
\arguments{
\item{light_threshold}{}
}
\description{
Title
}
