% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_from_vector.R
\name{spectrace_from_vector}
\alias{spectrace_from_vector}
\title{Unpack vector spectral data into columns}
\usage{
spectrace_from_vector(lightData)
}
\arguments{
\item{lightData}{Data frame containing the light data. Column with vectors of
spectral data must be named "spectrum" and with wavelength data must be named
"wavelength".}
}
\value{
The data frame with the spectral data with one column per wavelength.
}
\description{
This function unpacks spectral data stored in a vector into wavelength columns.
}
