% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import_activity.R
\name{spectrace_import_activity}
\alias{spectrace_import_activity}
\title{Import raw spectrace activity data}
\usage{
spectrace_import_activity(actFile, tz, serial_number = NA)
}
\arguments{
\item{actFile}{Path to the file containing the activity data.}

\item{tz}{Time zone to be used for datetime conversion. See supported time zones
by calling \code{\link[base]{OlsonNames}}.}

\item{serial_number}{Serial number of spectrace device. Defaults to NULL.}
}
\value{
Data frame.
}
\description{
This function imports the raw activity data of the spectrace sensors. It assumes
that the data is stored in CSV format.
}
