% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import_aopic.R
\name{spectrace_import_aopic}
\alias{spectrace_import_aopic}
\title{Import raw spectrace data, calibrate data, and calculate alpha-opic quantities.}
\usage{
spectrace_import_aopic(lightFile, tz, serial_number = NULL, cal_data = NULL)
}
\arguments{
\item{lightFile}{Path to the file containing the light data.}

\item{tz}{Time zone to be used for datetime conversion. See supported time zones
by calling \code{\link[base]{OlsonNames}}.}

\item{serial_number}{Serial number of spectrace device. Defaults to NULL.}

\item{cal_data}{Data frame with calibration factors. Must consist of the
following columns: "serial", "lux", "410nm", "435nm", "460nm", "485nm",
"510nm", "535nm","560nm", "585nm", "610nm", "645nm", "680nm", "705nm",
"730nm", "760nm". If not provided, the stored calibration data
will be used (default).}
}
\value{
Data frame.
}
\description{
This function imports raw spectrace output and calculates alpha-opic quantities.
See \code{\link{spectrace_import_light}},
\code{\link{spectrace_calibrate_light}} and \code{\link{spectrace_aopic}}
for more information.
}
