% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import_battery.R
\name{spectrace_import_battery}
\alias{spectrace_import_battery}
\title{Import raw spectrace battery data}
\usage{
spectrace_import_battery(batteryFile, tz)
}
\arguments{
\item{batteryFile}{Path to the file containing the battery data.}

\item{tz}{Time zone to be used for datetime conversion. See supported time zones
by calling \code{\link[base]{OlsonNames}}.}
}
\value{
Data frame.
}
\description{
This function imports the raw battery data of the spectrace sensors. It assumes
that the data is stored in CSV format.
}
