% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import_light.R
\name{spectrace_import_light}
\alias{spectrace_import_light}
\title{Import raw spectrace light data}
\usage{
spectrace_import_light(lightFile, tz)
}
\arguments{
\item{lightFile}{Path to the file containing the light data.}

\item{tz}{Time zone to be used for datetime conversion. See supported time zones
by calling \code{\link[base]{OlsonNames}}.}
}
\value{
Data frame.
}
\description{
This function imports the raw light data of the spectrace sensors. It assumes
that the data is stored in CSV format.
}
