% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_interpolate_spectra.R
\name{spectrace_interpolate_spectra}
\alias{spectrace_interpolate_spectra}
\title{Interpolate spectral data}
\usage{
spectrace_interpolate_spectra(
  lightData,
  resolution = c("5nm", "1nm"),
  interp_method = c("pchip", "linear")
)
}
\arguments{
\item{lightData}{Data frame containing the calibrated light data
for the channels from 410nm to 730nm.}

\item{resolution}{Integer. Resolution of the}

\item{interp_method}{Method for interpolation. Can be "pchip" (default) or
"linear".}
}
\value{
Data frame with illuminance, alpha-opic irradiances, alpha-opic EDI,
alpha-opic ELR, alpha-opic DER, and CCT.
}
\description{
This function interpolates Spectrace's spectral irradiance output to 5nm
resolution.
}
