% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_normalize_spectra.R
\name{spectrace_normalize_spectra}
\alias{spectrace_normalize_spectra}
\title{Normalize spectrace data}
\usage{
spectrace_normalize_spectra(
  lightData,
  method = c("peak", "AUC", "wavelength", "response"),
  wavelength = NULL,
  response = c("vl", "sc", "mc", "lc", "rod", "mel"),
  keepNormCoefficient = FALSE
)
}
\arguments{
\item{lightData}{Data frame containing the (calibrated) light data. Data needs
to be in wide format (see \code{\link{spectrace_to_wide}}).}

\item{method}{String specifying the normalization method. If method is "peak"
(default), data is normalized such that the peak (max value) is equal to 1.
If method is "AUC", data is normalized such that the area under the curve
is equal to 1. If method is "wavelength", data is normalized to 1 at specified
wavelength. If method is "response", data is normalized such that the response
is one (e.g., if response is "vl", the data is normalized such that the data is
normalized to vlambda).}

\item{wavelength}{Numeric. Wavelength to normalize at. Must be specified if
method is "wavelength". Defaults to NULL.}

\item{response}{String. Spectral response function to normalize at. Must be one of
["vl","sc", "mc, "lc", "rod", "mel"]. Defaults to "vl".}

\item{keepNormCoefficient}{Logical. Should the normalization coefficient be
kept in the data? Defaults to FALSE}
}
\value{
The original data frame with the spectral data replaced by the
   normalized spectral data.
}
\description{
This function normalizes the spectral data.
}
