% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_plot_clusters.R
\name{spectrace_plot_clusters}
\alias{spectrace_plot_clusters}
\title{Plot clustered spectrace data}
\usage{
spectrace_plot_clusters(
  lightData,
  lightData.PCA = NULL,
  sil.scores = NULL,
  samplesize = NULL
)
}
\arguments{
\item{lightData}{Data frame containing the (calibrated) light data. Data needs
to be in wide format (see \code{\link{spectrace_to_wide}}). Must contain a
`cluster_id` column.}

\item{lightData.PCA}{(Optional) data frame containing the principal components
after PCA of the lightData. Must contain a `cluster_id` column. Defaults to NULL.}

\item{sil.scores}{(Optional) data frame with silhouette scores per cluster.
Must contain a `cluster_id` and `sil_score` column. Defaults to NULL.}

\item{samplesize}{(Optional) Integer, indicating the size of the random sample of spectra
plotted per cluster. Defaults to NULL.}
}
\description{
Plot clustered spectrace data
}
