% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_to_long.R
\name{spectrace_to_long}
\alias{spectrace_to_long}
\title{Spectral data to long format}
\usage{
spectrace_to_long(lightData)
}
\arguments{
\item{lightData}{Data frame with light data. Must contain one column per
wavelength, named '[wavelength]nm' (e.g., '380nm').}
}
\value{
Reshaped data.
}
\description{
Reshapes spectral data to long format, with one variable for the wavelength and
one for the power per wavelength.
}
