% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_to_vector.R
\name{spectrace_to_vector}
\alias{spectrace_to_vector}
\title{Pack spectral data into vectors}
\usage{
spectrace_to_vector(lightData)
}
\arguments{
\item{lightData}{Data frame containing the light data. Column names for
spectral data must be the wavelength followed by "nm", for example: "380nm".}
}
\value{
The data frame with the spectrum and wavelength as vectors.
}
\description{
This function packs spectral data into a vector.
}
