% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_to_wide.R
\name{spectrace_to_wide}
\alias{spectrace_to_wide}
\title{Spectral data to wide format}
\usage{
spectrace_to_wide(lightData)
}
\arguments{
\item{lightData}{Data frame with light data. Must contain one column with the
wavelengths named "wl" and one column with the values per wavelength named
"val".}
}
\value{
Reshaped data.
}
\description{
Reshapes spectral data from long to wide format, with one column per wavelength.
}
