% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_aopic.R
\name{spectrace_aopic}
\alias{spectrace_aopic}
\title{Calculate alpha-opic quantities from raw spectrace data}
\usage{
spectrace_aopic(lightData)
}
\arguments{
\item{lightData}{Data frame containing the calibrated light data
for the channels from 410nm to 730nm.}
}
\value{
Data frame with illuminance, alpha-opic irradiances, alpha-opic EDI,
alpha-opic ELR, alpha-opic DER, and CCT.
}
\description{
This function calculates photopic and alpha-opic quantities as defined in the
CIE s26e standard, from the raw spectrace data. Before calculations, the raw
spectrace data is converted to spectral irradiance using provided or default
calibration data. Spectral irradiance is then interpolated to 5nm resolution
using piecewise cubic hermitean interpolation polynomials. CCT is calculated
with McCamy's approximation.
}
