% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_aopic.R
\name{spectrace_aopic}
\alias{spectrace_aopic}
\title{Calculate alpha-opic quantities from calibrated spectrace data}
\usage{
spectrace_aopic(
  lightData,
  interp_method = c("pchip", "linear", "none"),
  keep_spectral_data = TRUE
)
}
\arguments{
\item{lightData}{Data frame containing the calibrated light data.x = data.}

\item{interp_method}{Method for interpolation. Can be "pchip" (smooth
piecewise hermetic interpolation) or "linear". Defaults to "pchip".}

\item{keep_spectral_data}{Logical. Should the spectral irradiance columns be kept?
Defaults to TRUE.}
}
\value{
Data frame extended with values for illuminance, alpha-opic irradiances,
   alpha-opic EDI, alpha-opic ELR, alpha-opic DER, and CCT.
   If \code{keep_spectral_data} is FALSE then the spectral data columns will be
   removed from the original data frame.
}
\description{
This function calculates photopic and alpha-opic quantities as defined in the
CIE s26e standard, from the calibrated spectrace data. Spectral irradiance is
interpolated to 5nm resolution. CCT is calculated
with McCamy's approximation.
}
