% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import.R,
%   R/spectrace_import_activity.R
\name{spectrace_import}
\alias{spectrace_import}
\title{Import raw spectrace data}
\usage{
spectrace_import(actFile, tz)

spectrace_import(actFile, tz)
}
\arguments{
\item{actFile}{Path to the file containing the activity data.}

\item{tz}{Time zone to be used for datetime conversion. See supported time zones
by calling \code{\link[base]{OlsonNames}}.}

\item{lightFile}{Path to the file containing the light data.}

\item{offset}{Time offset between activity and light data in secs (activity - light).}
}
\value{
Data frame.

Data frame.
}
\description{
This function imports the raw output data of the spectrace sensors. It assumes
that the data is stored in CSV format. If activity data is provided, these data
will be merged with the light data.

This function imports the raw activity data of the spectrace sensors. It assumes
that the data is stored in CSV format.
}
