% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import_aopic.R
\name{spectrace_import_aopic}
\alias{spectrace_import_aopic}
\title{Import raw spectrace data and calculate alpha-opic quantities.}
\usage{
spectrace_import_aopic(
  lightFile,
  actFile = NULL,
  offset = 0,
  tz,
  cal = NULL,
  include_raw = FALSE
)
}
\arguments{
\item{lightFile}{Path to the file containing the light data.}

\item{actFile}{Path to the file containing the activity data. Defaults to NULL.}

\item{offset}{Time offset between activity and light data in secs (activity - light).}

\item{tz}{Time zone to be used for datetime conversion.}

\item{cal}{Calibration data for the spectrace irradiance sensors. If no data is
provided (default), default calibration data is used (not recommended).}

\item{include_raw}{Logical. Should the raw irradiance output be included?
Defaults to FALSE.}
}
\value{
Data frame.
}
\description{
This function imports raw spectrace output and calculates alpha-opic quantities.
See \code{\link{spectrace_import}} and \code{\link{spectrace_aopic}}
for more information.
}
