% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_import_combined.R
\name{spectrace_import_combined}
\alias{spectrace_import_combined}
\title{Title}
\usage{
spectrace_import_combined(
  lightFile,
  activityFile = NA,
  batteryFile = NA,
  tz = NA,
  serial_number = NA
)
}
\arguments{
\item{serial_number}{}
}
\description{
Title
}
