% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrace_interpolate_spectra.R
\name{spectrace_interpolate_spectra}
\alias{spectrace_interpolate_spectra}
\title{Interpolate spectral data}
\usage{
spectrace_interpolate_spectra(
  lightData,
  resolution = c("5nm", "1nm"),
  interp_method = c("pchip", "linear"),
  normalize = FALSE
)
}
\arguments{
\item{lightData}{Data frame containing the calibrated light data
for the channels from 410nm to 730nm. Additional variables are allowed.}

\item{resolution}{String specifying the resolution of the output
spectrum. Can be "5nm" (default) or "1nm".}

\item{interp_method}{The interpolation method. Can be either "pchip" (default)
or "linear". Pchip (piecewise cubic hermetic interpolation) results in a
smooth spectrum while preserving the source values as local minima/maxima.}

\item{normalize}{Logical. Should the interpolated spectrum be normalized to
peak = 1?}
}
\value{
The original data frame with the spectral data replaced by the
   interpolated spectral data.
}
\description{
This function interpolates Spectrace's spectral irradiance output to visible
range (380-780nm) with 1nm or 5nm resolution.
}
